pyproject.toml
[project]
name = "story-teller-with-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Farhaj499", email = "farhajshauket@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.102.0",
    "crewai-tools>=0.36.0",
    "langchain-google-genai>=2.0.9",
    "litellm>=1.60.2",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
story-teller = "story_teller_with_crewai.main:main"


